API_TUTORIAL/requirements.txt
asgiref==3.6.0
beautifulsoup4==4.12.0
certifi==2022.12.7
chardet==3.0.4
charset-normalizer==3.1.0
coreapi==2.3.3
coreschema==0.0.4
crispy-bootstrap5==0.7
Django==4.1.6
django-appconf==1.0.5
django-autocomplete-light==3.9.4
django-ckeditor==6.5.1
django-crispy-forms==2.0
django-crontab==0.7.1
django-filter==22.1
django-js-asset==2.0.0
django-modeltranslation==0.18.9
django-phonenumber-field==7.0.2
django-select2==8.1.0
djangorestframework==3.14.0
drf-yasg==1.21.5
googletrans==4.0.0rc1
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2023.1.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.2
packaging==23.1
phonenumberslite==8.13.6
Pillow==9.4.0
pytz==2023.3
requests==2.31.0
rfc3986==1.5.0
ruamel.yaml==0.17.28
ruamel.yaml.clib==0.2.7
six==1.16.0
sniffio==1.3.0
soupsieve==2.4
sqlparse==0.4.3
typing_extensions==4.5.0
tzdata==2022.7
uritemplate==4.1.1
urllib3==1.26.15
